requirements.txt
langchain
langchain_core
langchain_community
pypdf
python-dotenv
sentence-transformers
chromadb
langchain-huggingface